+++ ibllib/atlas/atlas.py
@@ -366,6 +366,22 @@ class BrainAtlas:
         ax.imshow(im, extent=extent, cmap=cmap, **kwargs)
         return ax
 
+    def extent(self, axis):
+        """
+        :param axis: direction along which the volume is stacked:
+         (2 = z for horizontal slice)
+         (1 = y for coronal slice)
+         (0 = x for sagittal slice)
+        :return:
+        """
+        if axis == 0:
+            extent = np.r_[self.bc.ylim, np.flip(self.bc.zlim)] * 1e6
+        elif axis == 1:
+            extent = np.r_[self.bc.xlim, np.flip(self.bc.zlim)] * 1e6
+        elif axis == 2:
+            extent = np.r_[self.bc.xlim, np.flip(self.bc.ylim)] * 1e6
+        return extent
+
     def slice(self, coordinate, axis, volume='image', mode='raise', region_values=None,
               mapping="Allen"):
         """
@@ -413,38 +429,35 @@ class BrainAtlas:
         elif volume in ['surface', 'edges']:
             return _take(self.surface, index, axis=self.xyz2dims[axis])
 
-    def plot_cslice(self, ap_coordinate, volume='image', **kwargs):
+    def plot_cslice(self, ap_coordinate, volume='image', mapping='Allen', **kwargs):
         """
         Imshow a coronal slice
         :param: ap_coordinate (m)
         :param volume: 'image' or 'annotation'
         :return: ax
         """
-        cslice = self.slice(ap_coordinate, axis=1, volume=volume)
-        extent = np.r_[self.bc.xlim, np.flip(self.bc.zlim)] * 1e6
-        return self._plot_slice(cslice.T, extent=extent, **kwargs)
+        cslice = self.slice(ap_coordinate, axis=1, volume=volume, mapping=mapping)
+        return self._plot_slice(cslice.T, extent=self.extent(axis=1), **kwargs)
 
-    def plot_hslice(self, dv_coordinate, volume='image', **kwargs):
+    def plot_hslice(self, dv_coordinate, volume='image', mapping='Allen', **kwargs):
         """
         Imshow a horizontal slice
         :param: dv_coordinate (m)
         :param volume: 'image' or 'annotation'
         :return: ax
         """
-        hslice = self.slice(dv_coordinate, axis=2, volume=volume)
-        extent = np.r_[self.bc.xlim, np.flip(self.bc.ylim)] * 1e6
-        return self._plot_slice(hslice, extent=extent, **kwargs)
+        hslice = self.slice(dv_coordinate, axis=2, volume=volume, mapping=mapping)
+        return self._plot_slice(hslice, extent=self.extent(axis=2), **kwargs)
 
-    def plot_sslice(self, ml_coordinate, volume='image', **kwargs):
+    def plot_sslice(self, ml_coordinate, volume='image', mapping='Allen', **kwargs):
         """
         Imshow a sagittal slice
         :param: ml_coordinate (m)
         :param volume: 'image' or 'annotation'
         :return: ax
         """
-        sslice = self.slice(ml_coordinate, axis=0, volume=volume)
-        extent = np.r_[self.bc.ylim, np.flip(self.bc.zlim)] * 1e6
-        return self._plot_slice(np.swapaxes(sslice, 0, 1), extent=extent, **kwargs)
+        sslice = self.slice(ml_coordinate, axis=0, volume=volume, mapping=mapping)
+        return self._plot_slice(np.swapaxes(sslice, 0, 1), extent=self.extent(axis=0), **kwargs)
 
     def plot_top(self, ax=None):
         ix, iy = np.meshgrid(np.arange(self.bc.nx), np.arange(self.bc.ny))
@@ -454,7 +467,7 @@ class BrainAtlas:
             ax = plt.gca()
             ax.axis('equal')
         ax.imshow(self._label2rgb(self.label.flat[inds]),
-                  extent=np.r_[self.bc.xlim, np.flip(self.bc.ylim)] * 1e6, origin='upper')
+                  extent=self.extent(axis=2), origin='upper')
         return ax
 
 
@@ -711,7 +724,14 @@ class AllenAtlas(BrainAtlas):
         """
         par = params.read('one_params')
         FLAT_IRON_ATLAS_REL_PATH = Path('histology', 'ATLAS', 'Needles', 'Allen')
+        LUT_VERSION = "v01"  # version 01 is the lateralized version
         regions = BrainRegions()
+        xyz2dims = np.array([1, 0, 2])  # this is the c-contiguous ordering
+        dims2xyz = np.array([1, 0, 2])
+        # we use Bregma as the origin
+        self.res_um = res_um
+        ibregma = (ALLEN_CCF_LANDMARKS_MLAPDV_UM['bregma'] / self.res_um)
+        dxyz = self.res_um * 1e-6 * np.array([1, -1, -1]) * scaling
         if mock:
             image, label = [np.zeros((528, 456, 320), dtype=np.int16) for _ in range(2)]
             label[:, :, 100:105] = 1327  # lookup index for retina, id 304325711 (no id 1327)
@@ -725,22 +745,23 @@ class AllenAtlas(BrainAtlas):
             file_label = path_atlas.joinpath(f'annotation_{res_um}.nrrd')
             if not file_label.exists():
                 _download_atlas_flatiron(file_label, FLAT_IRON_ATLAS_REL_PATH, par)
-            file_label_remap = path_atlas.joinpath(f'annotation_{res_um}_lut.npz')
+            file_label_remap = path_atlas.joinpath(f'annotation_{res_um}_lut_{LUT_VERSION}.npz')
             if not file_label_remap.exists():
                 label = self._read_volume(file_label)
                 _logger.info("computing brain atlas annotations lookup table")
+                # lateralize atlas: for this the regions of the left hemisphere have primary
+                # keys opposite to to the normal ones
+                lateral = np.zeros(label.shape[xyz2dims[0]])
+                lateral[int(np.floor(ibregma[0]))] = 1
+                lateral = np.sign(np.cumsum(lateral)[np.newaxis, :, np.newaxis] - 0.5)
+                label = label * lateral
                 _, im = ismember(label, regions.id)
                 label = np.reshape(im.astype(np.uint16), label.shape)
+                _logger.info(f"saving {file_label_remap} ...")
                 np.savez_compressed(file_label_remap, label)
             # loads the files
             label = self._read_volume(file_label_remap)
             image = self._read_volume(file_image)
-        xyz2dims = np.array([1, 0, 2])  # this is the c-contiguous ordering
-        dims2xyz = np.array([1, 0, 2])
-        dxyz = res_um * 1e-6 * np.array([1, -1, -1]) * scaling
-        # we use Bregma as the origin
-        ibregma = (ALLEN_CCF_LANDMARKS_MLAPDV_UM['bregma'] / res_um)
-        self.res_um = res_um
         super().__init__(image, label, dxyz, regions, ibregma,
                          dims2xyz=dims2xyz, xyz2dims=xyz2dims)